+++ tests/Events/EventContextTest.php
@@ -280,11 +280,11 @@ final class EventContextTest extends TestCase
 
         $coreRef = new \ReflectionClass(SharedFeatureflipCore::class);
         $core = $coreRef->newInstanceWithoutConstructor();
-        $coreRef->getConstructor()->invoke($core, $store, $this->processor, null, null, []);
+        $coreRef->getMethod('__construct')->invoke($core, $store, $this->processor, null, null, []);
 
         $clientRef = new \ReflectionClass(FeatureflipClient::class);
         $client = $clientRef->newInstanceWithoutConstructor();
-        $clientRef->getConstructor()->invoke($client, $core);
+        $clientRef->getMethod('__construct')->invoke($client, $core);
 
         return $client;
     }